refactor: moved getProposalLabel to utils/proposals and renamed to getGuildProposalLabel
