+++ apps/davi/src/components/ProposalVoteCard/components/VoteResults/VoteResults.tsx
@@ -5,7 +5,7 @@ import { Loading } from 'components/primitives/Loading';
 import { useTheme } from 'styled-components';
 import { useTranslation } from 'react-i18next';
 
-import { getOptionLabel } from 'utils/guildsProposals';
+import { getGuildOptionLabel } from 'utils/proposals';
 import {
   ResultRowProps,
   VoteResultsProps,
@@ -32,7 +32,11 @@ const VoteResultRow: React.FC<ResultRowProps> = ({
     voteData?.totalLocked,
     2
   );
-  const label = getOptionLabel({ metadata: proposalMetadata, optionKey, t });
+  const label = getGuildOptionLabel({
+    metadata: proposalMetadata,
+    optionKey,
+    t,
+  });
   return (
     <VotesRowWrapper>
       <VoteOption>
