+++ apps/davi/src/utils/proposals.test.ts
@@ -1,4 +1,4 @@
-import { getOptionLabel } from './guildsProposals';
+import { getGuildOptionLabel } from './proposals';
 
 const mockMetadataWithData = {
   voteOptions: ['Against', 'First option', 'Second option'],
@@ -11,10 +11,10 @@ const mockMetadataEmpty = {
 const mockT = key => key;
 
 describe('guildsProposals', () => {
-  describe('getOptionLabel', () => {
+  describe('getGuildOptionLabel', () => {
     it('should return the label for an option with label', () => {
       mockMetadataWithData.voteOptions.forEach((option, index) => {
-        const result = getOptionLabel({
+        const result = getGuildOptionLabel({
           metadata: mockMetadataWithData,
           optionKey: index,
           t: mockT,
@@ -25,7 +25,7 @@ describe('guildsProposals', () => {
 
     it('without labels, it should return "against" for the first option and "option" for the others', () => {
       mockMetadataWithData.voteOptions.forEach((option, index) => {
-        const result = getOptionLabel({
+        const result = getGuildOptionLabel({
           metadata: mockMetadataEmpty,
           optionKey: index,
           t: mockT,
