+++ apps/davi/src/utils/proposals.ts
@@ -438,3 +438,23 @@ export const decodeProposalStatus = function (
       };
   }
 };
+
+export function mapStructToProposal(
+  proposalStruct: any,
+  proposalId: string
+): Proposal {
+  const proposal: Proposal = {
+    ...proposalStruct,
+    id: proposalId,
+  };
+  return proposal;
+}
+
+export const getGuildOptionLabel = ({ metadata, optionKey, t }) => {
+  const metadataLabel = metadata?.voteOptions?.[optionKey];
+  return metadataLabel
+    ? metadataLabel
+    : Number(optionKey) === 0
+    ? t('actionBuilder.options.against', { defaultValue: 'Against' })
+    : t('actionBuilder.options.option', { optionKey });
+};
